Cross-Compare Temporal Boundaries

Given a four-element vector of start and end coordinates of two events, cross.time() compares the distances among the upper and lower boundaries of pairs of event vectors. This includes "like" boundary comparison (e.g., start #2 - start#1) and contrary boundary comparison (e.g. start #2 - end #1).

cross.time() memoises arguments s0, s1, e0, and e1 with a disk cache defined as cache_disk(dir = tempdir(), max_age = Inf, max_n = Inf, destroy_on_finalize = FALSE). A future version will allow the cache to be customized in a post-hoc manner.

Usage

cross.time(s0, s1, e0, e1, control = list(-Inf, Inf), chatty = FALSE, ...)

Arguments

s0

A numeric/date-coded vector containing the temporal lower boundary of the starting event duration

s1

A numeric/date-coded vector containing the temporal upper boundary of the starting event duration

e0

A numeric/date-coded vector containing the temporal lower boundary of the ending event duration

e1

A numeric/date-coded vector containing the temporal upper boundary of the ending event duration

control

A length-2 sorted list with values indicating the range of allowable values for internal variable beta (the difference between ends of 'to' events and beginnings of 'from' events)

chatty

(logical | FALSE) Verbosity flag

...

(Not used)

Value

A data.table object having the following fields:

mGap:

Metric describing the difference between the following temporal boundaries: TO.start, FROM.end

mSt:

Metric describing the difference between the following temporal boundaries: TO.start, FROM.start

mEd:

Metrics describing the difference between the following temporal boundaries: TO.end, FROM.end

from.len:

The duration of time of each "from" event: units are one the scale of the smallest increment of time represented (e.g., calendar dates in days will have lengths expressed in days)

to.len:

The duration of time of each "to" event: units are one the scale of the smallest increment of time represented (e.g., calendar dates in days will have lengths expressed in days)

epsilon:

A complex number (e.g., 1337 + 0.90210i) describing the relational changes from one event to another with interpretation based on whether or not the real and imaginary parts are > 0, < 0, or == 0:

Re	Im	Desc

> 0	0	Disjoint

0	> 0	Concurrency

> 0	> 0	Full Concurrency

0	0	Continuity

< 1	to event shorter than from event

= 1	to event same length as from event

> 1	to event longer than from event

epsilon.desc:

A plain-language description of epsilon