Relational Temporal Vector Space

`event.vector.space` is an R6 class object that creates a temporally-compared space of "event vectors", each comprised of a 'from' and 'to' date.

The time between these events is the focus of derivation: properly, the event vector is a complex number that encodes the relationship between the boundaries of the events, thus allowing one to describe this relationship in a concise manner.

Execution Workflow

The initial execution order should look something like the following ...
event.vector.space$new(...)$configure(...)$set.data(...)$set.q_graphs(...) %>% make.evs_universe(...) . The ability to execute the preceding workflow out of order exists, but it is best to adhere to the provided flow the first time around.

Class Member "Space"

$space is a data.table that is populated upon execution of make.evs_universe():

jk

Values of the "join key"

"crossed" time output

See cross.time()

fr_*, to_*

The temporal ranges of each from/toevent

src.pair

Values of the "from-to" pairings of events given by the contexts provided: FROM:TO

from.coord

String representation of temporal boundaries expressed as concatenated integers

to.coord

String representation of temporal boundaries expressed as concatenated integers

Public fields

space

The object containing the derived set of data (see 'Details')

q_graph

A 'query graph' list containing the event interactions for each level of 'jk' (see 'Details')

evt_graphs

Event graphs for each level of 'jk' (see 'Details') Initialize a Vector Space

Active bindings

config

This is an active binding that returns the configuration used to instantiate the class.

Methods

Public methods

event.vector.space$new()

event.vector.space$configure()

event.vector.space$set.data()

event.vector.space$set.q_graphs()

event.vector.space$clone()

Method new()

Usage

event.vector.space$new(config = NULL, ...)

Arguments

config

(see class method $configure())

...

(not used)

events.ascending

When TRUE, the final output will only contain chronological events.

Method configure()

Configuration is achieved via the object passed to config used at initialization or this method:

If the former method is used, object must be a data.frame, data.table, list(-like) with the following columns: src.names, contexts, map.fields, and row.filters. It is important to set the configuration data.frame correctly by ensuring the elements of each row are related to the same dataset; otherwise, the wrong filter may attempted to be applied to a non-existent column.

If using this method, the same fields are available as arguments with the additional arguments src.mix and exclude.mix that can be used to override the default behavior of how class method $evs.universe() cross-compares sources.

Usage

event.vector.space$configure(
  src.names,
  contexts,
  map.fields,
  row.filters,
  src.mix = "comb",
  exclude.mix = list(c("", "")),
  update = FALSE,
  chatty = FALSE
)

Arguments

src.names

An atomic vector of strings containing the R workspace names of the source datasets given in "ENVIRONMENT$object" or "object" format

contexts

An atomic vector of strings serving as labels for each data source

map.fields

One of two forms are supported:

An atomic vector of delimited strings, each string indicating the columns that serve as the join-key (jk), lower boundary (time_start_idx) and upper boundary (time_end_idx) in that order

A list of length-three array of strings following the ordering rule stated above

row.filters

One of two forms are supported:

An atomic vector of strings, each element being used to filter the corresponding dataset before processing. Each string should contain a logical statement of predicates such as \((a \le b) | ((i + 10) \ge k[1])\)

An expression list of similar construction

src.mix

"combination":

The default: generates unique pairs of sources

"reflexive":

Compares each source to itself

"all":

The union of the preceding options

Partial matching is allowed, and generated combinations include the mirror (e.g., "A, B" will have a "B, A" combo pair generated).

exclude.mix

A list of vectors containing each source pair context to exclude (e.g. list(c("A", "C"), c("u", "k"))). evs_exclude.blender can be invoked to create this list more quickly. Combinations are not automatically mirrored as is the case with 'src.mix'.

update

(logical) TRUE results in appending to the existing configuration

chatty

(logical | FALSE) Verbosity flag

Method set.data()

set.data() adds column src to the objects referenced by the configuration argument

Usage

event.vector.space$set.data(chatty = FALSE)

Arguments

chatty

(logical | FALSE) Verbosity flag

Method set.q_graphs()

set.q_graphs() creates a list of 'query graphs' for each unique value of 'jk'. The list is stored as class member q_graph

Usage

event.vector.space$set.q_graphs(chatty = FALSE)

Arguments

chatty

(logical | FALSE) Verbosity flag

Method clone()

The objects of this class are cloneable with this method.

Usage

event.vector.space$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.